try to bring it into HomeAccessory class
